rework flex formatting
The size of the grid was defined in terms of the size of the squares.
This was a problem because a later stage of the project requires a
dynamic number of squares, with a constant grid pixel size.
Now the size of the squares is defined by the size of the grid, which
is constant at 800x800px